Use hashes for hassio entrypoints (#6260)
